MicrosoftCalendar-SAP-Timesheet-App/requirements.txt
streamlit==1.39.0
streamlit_modal==0.1.2
streamlit_ace==0.1.1
fastapi==0.115.0
uvicorn==0.31.1
langchain==0.3.3
langchain_core==0.3.10
langchain_openai==0.2.2
openai==1.51.2
pdfplumber==0.11.4
requests==2.32.3
langchain_community==0.3.2
psycopg2_binary==2.9.9
sqlalchemy==2.0.35
loguru==0.7.2
tqdm==4.66.5
pandas==2.2.3
pydantic==2.9.2
fuzzywuzzy==0.18.0